requirements.txt
attrs==21.2.0 
asgiref==3.2.10            
certifi==2019.9.11     
chardet==3.0.4         
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==2.8                 
importlib-metadata==1.5.0
gunicorn==20.0.4
mixer==7.1.2  
more-itertools==8.2.0     
packaging==20.1           
pillow==8.3.1
pluggy==0.13.1 
psycopg2-binary==2.8.6
PyJWT==2.1.0               
pyparsing==2.4.6          
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4         
pytz==2020.1              
requests==2.26.0
six==1.14.0               
sorl-thumbnail==12.6.3
sqlparse==0.3.1           
urllib3==1.25.6           
wcwidth==0.1.8            
zipp==2.2.0               
django-email-users~=1.1.0
recipes~=0.1
djoser~=2.1.0
reportlab~=3.6.3
python-dotenv~=0.19.2